requirements.txt
fastapi 
uvicorn
cassandra-driver
python-dotenv
pydantic-settings